Move getTracks to lib module.
