+++ lib/index.js
@@ -1,3 +1,4 @@
 const getBody = require('./getBody');
+const getTracks = require('./getTracks');
 
-module.exports = { getBody };
+module.exports = { getBody, getTracks };
